# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VAKSINA HR - Kadrlarni boshqaruv dasturi
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Zayavka
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
    patch:
      operationId: updateRequest
      tags: [requests]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Arizani o'chirish (faqat HR va Direktor)
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "204":
          description: O'chirildi
        "403":
          description: Ruxsat yo'q
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      parameters:
        - { name: departmentId, in: query, schema: { type: number } }
        - { name: mentorId, in: query, schema: { type: number } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Xodimlar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Xodim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # --- VAZIFALAR (Tasks board) ---
...
components:
  schemas:
...
    Request:
      type: object
      required: [id, departmentId, position, count, status, priority, createdAt]
      properties:
        id: { type: number }
        departmentId: { type: number }
        departmentName: { type: ["string", "null"] }
        position: { type: string }
        count: { type: number }
        description: { type: ["string", "null"] }
        requirements: { type: ["string", "null"] }
        salaryRange: { type: ["string", "null"] }
        deadline: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        district: { type: ["string", "null"] }
        priority:
          type: string
          enum: [urgent, normal]
        status:
          type: string
          enum: [submitted, reviewing, accepted, announced, closed]
        assignedToId: { type: ["integer", "null"] }
        assignedToName: { type: ["string", "null"] }
        assignedAt: { type: ["string", "null"] }
        createdById: { type: ["integer", "null"] }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string }
        vacancyId: { type: ["integer", "null"] }
        vacancyCreatedAt: { type: ["string", "null"] }
        vacancyAssignedAt: { type: ["string", "null"] }
        vacancyAcceptedAt: { type: ["string", "null"] }
        vacancyPublishedAt: { type: ["string", "null"] }
...
    RequestUpdate:
      type: object
      properties:
        position: { type: string }
        count: { type: number }
        description: { type: string }
        requirements: { type: string }
        salaryRange: { type: string }
        deadline: { type: string }
        reason: { type: string }
        priority: { type: string }
        status: { type: string }
...
    Employee:
      type: object
      required: [id, fullName, position, departmentId, hiredAt, createdAt]
      properties:
        id: { type: number }
        fullName: { type: string }
        position: { type: string }
        departmentId: { type: number }
        departmentName: { type: ["string", "null"] }
        mentorId: { type: ["integer", "null"] }
        mentorName: { type: ["string", "null"] }
        hiredAt: { type: string }
        candidateId: { type: ["integer", "null"] }
        orgRole: { type: ["string", "null"], enum: [coordinator, manager, pharmacist, null] }
        reportsToId: { type: ["integer", "null"] }
        location: { type: ["string", "null"] }
        shiftType: { type: ["string", "null"], enum: [one, two, custom, null] }
        shiftLabel: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    EmployeeInput:
      type: object
      required: [fullName, position, departmentId, hiredAt]
      properties:
        fullName: { type: string }
        position: { type: string }
        departmentId: { type: number }
        mentorId: { type: number }
        hiredAt: { type: string }
        candidateId: { type: number }
        orgRole: { type: string, enum: [coordinator, manager, pharmacist] }
        reportsToId: { type: number }
        location: { type: string }
        shiftType: { type: string, enum: [one, two, custom] }
        shiftLabel: { type: string }
        photoUrl: { type: string }
    EmployeeUpdate:
      type: object
      properties:
        fullName: { type: string }
        position: { type: string }
        departmentId: { type: number }
        mentorId: { type: number }
        orgRole: { type: string, enum: [coordinator, manager, pharmacist] }
        reportsToId: { type: number }
        location: { type: string }
        shiftType: { type: string, enum: [one, two, custom] }
        shiftLabel: { type: string }
        photoUrl: { type: string }

    # --- NOTIFICATION ---
...